cleaning: move matchers tests in matcher folder
