+++ v-next/hardhat-chai-matchers/test/matchers/big-number.ts
@@ -3,7 +3,7 @@ import { describe, it } from "node:test";
 import { InvalidParameterError } from "@ignored/hardhat-vnext-utils/common-errors";
 import { AssertionError, expect } from "chai";
 
-import { addChaiMatchers } from "../src/internal/add-chai-matchers.js";
+import { addChaiMatchers } from "../../src/internal/add-chai-matchers.js";
 
 type SupportedNumber = number | bigint;
