+++ v-next/hardhat-chai-matchers/test/matchers/changeEtherBalances.ts
@@ -1,5 +1,5 @@
-import type { ChangeEtherBalance } from "./helpers/contracts.js";
-import type { Token } from "../src/internal/matchers/changeTokenBalance.js";
+import type { Token } from "../../src/internal/matchers/changeTokenBalance.js";
+import type { ChangeEtherBalance } from "../helpers/contracts.js";
 import type { EthereumProvider } from "@ignored/hardhat-vnext/types/providers";
 import type {
   HardhatEthers,
@@ -19,7 +19,7 @@ import {
 } from "@nomicfoundation/hardhat-test-utils";
 import { expect, AssertionError } from "chai";
 
-import { addChaiMatchers } from "../src/internal/add-chai-matchers.js";
+import { addChaiMatchers } from "../../src/internal/add-chai-matchers.js";
 
 addChaiMatchers();
 
@@ -528,7 +528,7 @@ describe("INTEGRATION: changeEtherBalances matcher", () => {
           ).to.changeEtherBalances(provider, [sender, receiver], [-100, 100]);
         } catch (e) {
           expect(util.inspect(e)).to.include(
-            path.join("test", "changeEtherBalances.ts"),
+            path.join("test", "matchers", "changeEtherBalances.ts"),
           );
           return;
         }
