+++ v-next/hardhat-chai-matchers/test/matchers/properAddress.ts
@@ -4,7 +4,7 @@ import { describe, it } from "node:test";
 
 import { expect, AssertionError } from "chai";
 
-import { addChaiMatchers } from "../src/internal/add-chai-matchers.js";
+import { addChaiMatchers } from "../../src/internal/add-chai-matchers.js";
 
 addChaiMatchers();
