+++ v-next/hardhat-chai-matchers/test/matchers/properHex.ts
@@ -2,7 +2,7 @@ import { describe, it } from "node:test";
 
 import { AssertionError, expect } from "chai";
 
-import { addChaiMatchers } from "../src/internal/add-chai-matchers.js";
+import { addChaiMatchers } from "../../src/internal/add-chai-matchers.js";
 
 addChaiMatchers();
