+++ v-next/hardhat-chai-matchers/test/matchers/properPrivateKey.ts
@@ -2,7 +2,7 @@ import { describe, it } from "node:test";
 
 import { expect, AssertionError } from "chai";
 
-import { addChaiMatchers } from "../src/internal/add-chai-matchers.js";
+import { addChaiMatchers } from "../../src/internal/add-chai-matchers.js";
 
 /* eslint-disable @typescript-eslint/no-unused-expressions -- allow all the expressions */
