+++ v-next/hardhat-chai-matchers/test/matchers/reverted/revertedWithCustomError.ts
@@ -1,4 +1,4 @@
-import type { MatchersContract } from "../helpers/contracts.js";
+import type { MatchersContract } from "../../helpers/contracts.js";
 import type { EthereumProvider } from "@ignored/hardhat-vnext/types/providers";
 import type { HardhatEthers } from "@ignored/hardhat-vnext-ethers/types";
 
@@ -15,13 +15,13 @@ import {
 } from "@nomicfoundation/hardhat-test-utils";
 import { AssertionError, expect } from "chai";
 
-import { addChaiMatchers } from "../../src/internal/add-chai-matchers.js";
-import { anyUint, anyValue } from "../../src/withArgs.js";
+import { addChaiMatchers } from "../../../src/internal/add-chai-matchers.js";
+import { anyUint, anyValue } from "../../../src/withArgs.js";
 import {
   runSuccessfulAsserts,
   runFailedAsserts,
   mineSuccessfulTransaction,
-} from "../helpers/helpers.js";
+} from "../../helpers/helpers.js";
 
 addChaiMatchers();
 
@@ -555,7 +555,12 @@ describe("INTEGRATION: Reverted with custom error", () => {
             "Expected transaction to be reverted with custom error 'SomeCustomError', but it reverted with reason 'some reason'",
           );
           expect(errorString).to.include(
-            path.join("test", "reverted", "revertedWithCustomError.ts"),
+            path.join(
+              "test",
+              "matchers",
+              "reverted",
+              "revertedWithCustomError.ts",
+            ),
           );
           return;
         }
