+++ v-next/hardhat-chai-matchers/test/matchers/reverted/revertedWithPanic.ts
@@ -1,4 +1,4 @@
-import type { MatchersContract } from "../helpers/contracts.js";
+import type { MatchersContract } from "../../helpers/contracts.js";
 import type { EthereumProvider } from "@ignored/hardhat-vnext/types/providers";
 import type { HardhatEthers } from "@ignored/hardhat-vnext-ethers/types";
 
@@ -15,13 +15,13 @@ import {
 } from "@nomicfoundation/hardhat-test-utils";
 import { AssertionError, expect } from "chai";
 
-import { addChaiMatchers } from "../../src/internal/add-chai-matchers.js";
-import { PANIC_CODES } from "../../src/panic.js";
+import { addChaiMatchers } from "../../../src/internal/add-chai-matchers.js";
+import { PANIC_CODES } from "../../../src/panic.js";
 import {
   runSuccessfulAsserts,
   runFailedAsserts,
   mineSuccessfulTransaction,
-} from "../helpers/helpers.js";
+} from "../../helpers/helpers.js";
 
 addChaiMatchers();
 
@@ -341,7 +341,7 @@ describe("INTEGRATION: Reverted with panic", () => {
             "Expected transaction NOT to be reverted with some panic code, but it reverted with panic code 0x01 (Assertion error)",
           );
           expect(errorString).to.include(
-            path.join("test", "reverted", "revertedWithPanic.ts"),
+            path.join("test", "matchers", "reverted", "revertedWithPanic.ts"),
           );
           return;
         }
